Move controllers to rails controller folder
